+++ packages/cloud-cognitive/src/components/StatusIcon/StatusIcon.test.js
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { render } from '@testing-library/react'; // https://testing-library.com/docs/react-testing-library/intro
+import { render, screen } from '@testing-library/react'; // https://testing-library.com/docs/react-testing-library/intro
 import React from 'react';
 
 import { pkg } from '../../settings';
@@ -16,8 +16,9 @@ import uuidv4 from '../../global/js/utils/uuidv4';
 import { StatusIcon } from '.';
 
 const blockClass = `${pkg.prefix}--status-icon`;
-const { componentName } = StatusIcon.displayName;
+const componentName = StatusIcon.displayName;
 const className = `class-${uuidv4()}`;
+const dataTestId = uuidv4();
 
 const iconSizes = [
   { input: 'sm', output: '16' },
@@ -61,7 +62,7 @@ describe(componentName, () => {
 
   it('has no accessibility violations', async () => {
     const { container } = renderComponent();
-    await expect(container).toBeAccessible(componentName);
+    await expect(container).toBeAccessible(componentName, 'scan_label');
     await expect(container).toHaveNoAxeViolations();
   }, 80000);
 
@@ -74,6 +75,11 @@ describe(componentName, () => {
     const ref = React.createRef();
     renderComponent({ ref });
     expect(ref.current).toHaveClass(blockClass);
+  });
+
+  it('adds additional properties to the containing node', () => {
+    renderComponent({ 'data-testid': dataTestId });
+    screen.getByTestId(dataTestId);
   });
 
   iconTypes.forEach((kind) => {
@@ -94,18 +100,18 @@ describe(componentName, () => {
     });
   });
 
-  iconTypes.forEach((desc) => {
-    it(`applies the proper title element when icon description of ${desc} is passed`, () => {
+  iconTypes.forEach((label) => {
+    it(`applies the proper title element when icon laebl of ${label} is passed`, () => {
       const { container } = render(
         <StatusIcon
-          kind={desc}
-          iconDescription={desc}
+          kind={label}
+          iconDescription={label}
           size="sm"
           theme="light"
         />
       );
       const element = container.querySelector(
-        `.${blockClass}--light.${blockClass}--light-${desc}`
+        `.${blockClass}--light.${blockClass}--light-${label}`
       );
       const hasIconDescriptionProp = element.querySelector('title').textContent;
       expect(hasIconDescriptionProp).toBeTruthy();