+++ packages/cloud-cognitive/src/components/index.js
@@ -8,6 +8,7 @@
 export { AboutModal } from './AboutModal';
 export { ActionBarItem } from './ActionBar';
 export { ContextHeader } from './ContextHeader';
+export { CreateModal } from './CreateModal';
 export {
   EmptyState,
   ErrorEmptyState,